Stage focus event
Stages no longer submit events to nodes when they aren't focused.

This will be changed in the future to exclude events that aren't a result of the user (so timers can still get through to Terminal nodes, which are a future node and thus the feature will be added with the node).