Stage adjust drawing
An option is now available 'Stage.noRedrawOnAdjust' which, when enabled will prevent the contents of the stage from being drawn while the stage is being moved or resized.

If this option is enabled the drawing will be blocked and during resize the stage will become blank.

If disabled the nodes will be drawn as usual.

If 'shift' is held while adjusting the stage the functionality is reversed. This means that if the setting is enabled and shift is held the content *will* be drawn.